+++ src/PowerShell/Authenticators/ServicePrincipalAuthenticator.cs
@@ -3,6 +3,7 @@
 
 namespace Microsoft.Store.PartnerCenter.PowerShell.Authenticators
 {
+    using Factories;
     using Identity.Client;
 
     /// <summary>
@@ -19,7 +20,31 @@ namespace Microsoft.Store.PartnerCenter.PowerShell.Authenticators
         /// </returns>
         public override AuthenticationResult Authenticate(AuthenticationParameters parameters)
         {
-            return null;
+            IConfidentialClientApplication app;
+            ServicePrincipalParameters servicePrincipalParameters = parameters as ServicePrincipalParameters;
+
+            if (string.IsNullOrEmpty(servicePrincipalParameters.Secret))
+            {
+                app = SharedTokenCacheClientFactory.CreateConfidentialClient(
+                    $"{parameters.Environment.ActiveDirectoryAuthority}{parameters.TenantId}",
+                    parameters.ApplicationId,
+                    null,
+                    GetCertificate(servicePrincipalParameters.CertificateThumbprint),
+                    null,
+                    parameters.TenantId);
+            }
+            else
+            {
+                app = SharedTokenCacheClientFactory.CreateConfidentialClient(
+                    $"{parameters.Environment.ActiveDirectoryAuthority}{parameters.TenantId}",
+                    parameters.ApplicationId,
+                    servicePrincipalParameters.Secret,
+                    null,
+                    null,
+                    parameters.TenantId);
+            }
+
+            return app.AcquireTokenForClient(parameters.Scopes).ExecuteAsync().ConfigureAwait(false).GetAwaiter().GetResult();
         }
 
         /// <summary>