+++ ChangeLog.md
@@ -23,9 +23,12 @@
 ## Upcoming Release
 
 * Authentication
+  * Added the ability to invoke [Connect-PartnerCenter](https://github.com/microsoft/Partner-Center-PowerShell/blob/master/docs/help/Connect-PartnerCenter.md) without requiring the creation of an Azure Active Directory application
   * Enabled interactive login support for cross-platform by default
   * Device code flow login is now the backup option of interactive login fails, or the user provides the `-UseDeviceAuthentication` switch parameter
   * Token cache is now shared with other products, such as Azure CLI and Visual Studio 2019
+* Module
+  * The `PartnerCenter` module now supports PowerShell 5.1 and PowerShell, as a result the `PartnerCenter.NetCore` module will be retired
 
 ## 1.5.1908.1